+++ examples/optimizer.jl
@@ -1,20 +1,20 @@
 #=====================================#
 # MIP solvers (commercial, but fast)  #
 #=====================================#
-function get_gurobi()
+function get_gurobi(; solver_log = true)
     GRB_ENV = Gurobi.Env()
     return optimizer_with_attributes(
         () -> Gurobi.Optimizer(GRB_ENV),
-        MOI.Silent() => true,
+        MOI.Silent() => !solver_log,
         # "MIPFocus" => 3, # Focus on optimality over feasibility 
         "Presolve" => 1,
     )
 end
 
-function get_cplex()
+function get_cplex(; solver_log = true)
     return JuMP.optimizer_with_attributes(
         CPLEX.Optimizer,
-        MOI.Silent() => true,
+        MOI.Silent() => !solver_log,
         # "CPX_PARAM_EPGAP" => 1E-6,
         # "CPX_PARAM_EPOPT" => 1E-8,
         # "CPX_PARAM_MIPEMPHASIS" => 1, # Focus on optimality over feasibility 
@@ -26,8 +26,8 @@ end
 # MIP solvers (open-source, but slow)  #
 #======================================#
 
-function get_glpk()
-    return JuMP.optimizer_with_attributes(GLPK.Optimizer, MOI.Silent() => false)
+function get_glpk(; solver_log = true)
+    return JuMP.optimizer_with_attributes(GLPK.Optimizer, MOI.Silent() => !solver_log)
 end
 
 # https://github.com/jump-dev/HiGHS.jl
@@ -43,10 +43,10 @@ end
 # Continuous nonlinear programming solver (open-source)  #
 #========================================================#
 
-function get_ipopt()
+function get_ipopt(; solver_log = true)
     return JuMP.optimizer_with_attributes(
         Ipopt.Optimizer,
-        MOI.Silent() => true,
+        MOI.Silent() => !solver_log,
         "sb" => "yes",
         "max_iter" => Int(1E4),
     )